% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnames.R
\name{cnames}
\alias{cnames}
\title{cnames()}
\usage{
cnames(
  name,
  toup = TRUE,
  rmcompanytxt = TRUE,
  blankcompany = TRUE,
  rm12letter = TRUE
)
}
\arguments{
\item{name}{string with name or name vector}

\item{toup}{required=TRUE converts names to uppercase}

\item{rmcompanytxt}{remove company characteristics from names}

\item{blankcompany}{convert names with company characteristics such as LLC, TRUST, etc. to BLANK}

\item{rm12letter}{if TRUE one and two letter words such as middle initial are removed}
}
\description{
clean names from a string
}
