% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlpubgoog.R
\name{dlpubgoog}
\alias{dlpubgoog}
\title{dlpubgoog}
\usage{
dlpubgoog(ss, sheet, destfile)
}
\arguments{
\item{ss}{ss of google sheet}

\item{sheet}{sheet}

\item{destfile}{destfile}
}
\description{
download a public google sheet link as csv
}
