% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exsubtxt.R
\name{exsubtxt}
\alias{exsubtxt}
\title{exsubtxt()
extract parts from a string based on defined string split}
\usage{
exsubtxt(string, nchar = 100, ps, firstonly = F)
}
\arguments{
\item{string}{string to split on to find what to extract}

\item{nchar}{number of characters from located string forward to extract}

\item{ps}{page source in text unlisted to find what to extract}

\item{firstonly}{if TRUE returns only first instance of extract}
}
\description{
exsubtxt()
extract parts from a string based on defined string split
}
