% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extxt.R
\name{extxt}
\alias{extxt}
\title{extxt()
extract parts from a string based on defined string split}
\usage{
extxt(string, nchar = 100, ps, firstonly = F)
}
\arguments{
\item{string}{string to split on to find what to extract}

\item{nchar}{number of characters from located string forward to extract}

\item{ps}{page source in text unlisted to find what to extract}

\item{firstonly}{if TRUE returns only first instance of extract}
}
\description{
extxt()
extract parts from a string based on defined string split
}
