% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extxt3.R
\name{extxt3}
\alias{extxt3}
\title{extxt3()
extract parts from a string based on defined string split}
\usage{
extxt3(string, ps, nchar = 100, firstonly = F)
}
\arguments{
\item{string}{string to split on to find what to extract}

\item{ps}{page source in text unlisted to find what to extract}

\item{nchar}{number of characters from located string forward to extract}

\item{firstonly}{if TRUE returns only first instance of extract}
}
\description{
extxt3()
extract parts from a string based on defined string split
}
