% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getnames.R
\name{getnames}
\alias{getnames}
\title{getnames()}
\usage{
getnames(namestring)
}
\arguments{
\item{namestring}{string with up to two names separated by '&', 'and' or '+'}
}
\description{
Get names from a string
}
