% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNCAATeam.R
\name{getNCAATeam}
\alias{getNCAATeam}
\title{getNCAATeam()}
\usage{
getNCAATeam(
  teamname
)
}
\arguments{
\item{teamname}{string name of team in any format}
}
\description{
Get unified team name and data from any format team name 
}
