% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTeam.R
\name{getTeam}
\alias{getTeam}
\title{getTeam()}
\usage{
getTeam(
  teamname, 
  out="teamonly", 
  upper=TRUE, 
  print=TRUE
)
}
\arguments{
\item{teamname}{string name of team in any format}

\item{out}{string 'teamonly' or 'to' for only team name or 'all' for data.frame of all team data}

\item{upper}{logical return team in upper case}

\item{print}{logical print team name in addition to returnK}
}
\description{
Get unified team name and data from any format team name 
}
