% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchwords.R
\name{matchwords}
\alias{matchwords}
\title{matchwords}
\usage{
matchwords(str1, str2)
}
\arguments{
\item{str1}{string or string vector 1}

\item{str2}{string or string vector 2}
}
\description{
Match of common words in two strings
}
