% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numcomwords.R
\name{numcomwords}
\alias{numcomwords}
\title{numcomwords}
\usage{
numcomwords(str1, str2)
}
\arguments{
\item{str1}{string1}

\item{str2}{string2}

\item{x}{data to recode from ASCII to latin}
}
\description{
number of common words between two strings
}
