% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numwords.R
\name{numwords}
\alias{numwords}
\title{numwords}
\usage{
numwords(str1, str2)
}
\arguments{
\item{str1}{string or string vector 1}

\item{str2}{string or string vector 2}
}
\description{
Count number of common words in two strings
}
