% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseaddress.R
\name{parseaddress}
\alias{parseaddress}
\title{parseaddress()}
\usage{
parseaddress(address, check_python = TRUE, force_stateabb = TRUE)
}
\arguments{
\item{address}{string of address to parse}

\item{check_python}{TRUE or FALSE, TRUE=do check for python dependencies and install if missing, FALSE is faster if you have already installed python3 and module usaddress}

\item{force_stateabb}{TRUE or FALSE, if TRUE state names are forced to abbreviation format for unified format}
}
\description{
Get address parts from a string address.
}
