% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recol.R
\name{recol}
\alias{recol}
\title{recol}
\usage{
recol(dataset, oldname, newname)
}
\arguments{
\item{dataset}{dataset}

\item{oldname}{what old col name}

\item{newname}{new col name}
}
\description{
rename column
}
