% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redoenc.R
\name{redoenc}
\alias{redoenc}
\title{redoenc}
\usage{
redoenc(x, fromenc = "ASCII", toenc = "latin1", sub = "", type = "vec")
}
\arguments{
\item{x}{data to recode from ASCII to latin}

\item{fromenc}{from encoding}

\item{toenc}{to encoding}

\item{type}{vec or df}
}
\description{
redo encoding
}
