% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yearExtract.R
\name{yearExtract}
\alias{yearExtract}
\title{yearExtract()
extract year from string}
\usage{
yearExtract(string)
}
\arguments{
\item{string}{string to extract from}
}
\description{
yearExtract()
extract year from string
}
